--- internal/server/crypto.go
@@ -1,43 +1,0 @@
-package server
-
-import (
-	"crypto/rand"
-	"crypto/rsa"
-	"crypto/sha256"
-	"crypto/x509"
-	"encoding/pem"
-	"fmt"
-	"os"
-)
-
-// parseCryptoPrivKey gets RSA private from file
-func parseCryptoPrivKey(path string) (*rsa.PrivateKey, error) {
-	fileBytes, err := os.ReadFile(path)
-	if err != nil {
-		return nil, err
-	}
-
-	block, _ := pem.Decode(fileBytes)
-	if block == nil {
-		return nil, fmt.Errorf("failed to parse PEM block containing the public key")
-	}
-
-	priv, err := x509.ParsePKCS1PrivateKey(block.Bytes)
-	if err != nil {
-		return nil, err
-	}
-
-	return priv, nil
-}
-
-// DecryptWithPrivateKey decrypts data with private key
-func decryptWithPrivateKey(ciphertext []byte, priv *rsa.PrivateKey) ([]byte, error) {
-	hash := sha256.New()
-
-	decryptedBytes, err := rsa.DecryptOAEP(hash, rand.Reader, priv, ciphertext, nil)
-	if err != nil {
-		return nil, err
-	}
-
-	return decryptedBytes, nil
-}
