--- internal/server/testdata/public.pem
@@ -1,9 +1,0 @@
------BEGIN PUBLIC KEY-----
-MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAt8nuYe84mTtq78aH8eIF
-98IRJUdKJ1N5UUt8DKkXGFVrDkMEjyp5kH8TGRar1uLdm9WYqkOKHJR6psA8BZKY
-nolap70wP5WcA62ujCXu9f0KAdrPalIOoO4Ez90qMs7MkanJvohgyeazAT1j2uaK
-rzey4lbf/uZhvAia61kqdzU6tGadSq15M/GCk6Tdym9f3Uf3ej7zXYnf5cz8CkVp
-kLAi5a6fHHTIguIKEtPysjSxB7k88T9NJqe8iqFEwv+Pk9wWnHAJFEOc5QRZ7PYR
-JQmoUM69ds9KdwFxiehq6l1e4GdKLqvlHUYc2zOzv9N1eSsWmSwNXbKX5GnJWY1U
-MQIDAQAB
------END PUBLIC KEY-----
